requirements.txt
# Networking + parsing
requests>=2.28
beautifulsoup4>=4.12
lxml>=4.9
tqdm>=4.65

# PDF / text
PyMuPDF>=1.22.3

# Embeddings / transformers
sentence-transformers>=2.2.2

# Vector DB - CPU
faiss-cpu==1.7.4

# Web API
fastapi>=0.95
uvicorn>=0.22

# utils
pandas>=2.0
